requirements.txt
PyYAML # to handle yaml files
beautifulsoup4
scrapy
requests
black # formatting the code

